@@ -1,14 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 28
     defaultConfig {
         applicationId "org.tinlone.demo.rxjavasample"
         minSdkVersion 19
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionCode 1
-        versionName "a.0"
+        versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
     // 编译选项
@@ -21,24 +20,6 @@ android {
     dexOptions {
         jumboMode = true
     }
-
-    // 签名配置
-//    signingConfigs {
-//        release {
-//            try {
-//                storeFile file("./keystore/release.keystore")
-//                storePassword "testres"
-//                keyAlias "testres"
-//                keyPassword "testres"
-//            } catch (ex) {
-//                throw new InvalidUserDataException(ex.toString())
-//            }
-//        }
-//
-//        debug {
-//            storeFile file("./keystore/debug.keystore")
-//        }
-//    }
 
     buildTypes {
         release {
@@ -58,29 +39,24 @@ android {
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-    compile project(path: ':vitamio')
-    compile files('libs/volley.jar')
-    compile 'com.android.support:design:26.1.0'
-    compile 'com.android.support:appcompat-v7:26.1.0'
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
-    compile 'com.squareup.retrofit2:retrofit:2.3.0'
-    compile 'com.squareup.retrofit2:converter-gson:2.2.0'
-    compile 'com.squareup.retrofit2:adapter-rxjava2:2.3.0'
-    compile 'io.reactivex.rxjava2:rxjava:2.1.7'
-    compile 'io.reactivex.rxjava2:rxandroid:2.0.1'
-    compile 'com.jakewharton:butterknife:8.6.0'
-    compile 'com.android.support:support-v4:26.1.0'
-    compile 'com.mylhyl:acp:1.1.7'
-    compile 'com.github.matecode:Snacky:1.0.2'
-    compile 'cat.ereza:customactivityoncrash:2.1.0'
-    compile 'com.tencent.bugly:crashreport_upgrade:1.3.4'
-    compile 'com.meituan.android.walle:library:1.1.5'
-    testCompile 'junit:junit:4.12'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation project(path: ':vitamio')
+    implementation 'com.android.support:design:28.0.0'
+    implementation 'com.android.support:appcompat-v7:28.0.0'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'com.squareup.retrofit2:retrofit:2.4.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
+    implementation 'com.squareup.retrofit2:adapter-rxjava2:2.4.0'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.3'
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.0'
+    implementation 'com.jakewharton:butterknife:8.6.0'
     annotationProcessor 'com.jakewharton:butterknife-compiler:8.6.0'
+    implementation 'com.android.support:support-v4:28.0.0'
+    implementation 'com.mylhyl:acp:1.1.7'
+    implementation 'com.github.matecode:Snacky:1.0.2'
+    implementation 'cat.ereza:customactivityoncrash:2.1.0'
+    implementation 'com.tencent.bugly:crashreport_upgrade:1.3.4'
+    implementation 'com.meituan.android.walle:library:1.1.5'
 }
 repositories {
     mavenCentral()
