@@ -6,7 +6,7 @@ import java.text.DecimalFormat;
 
 /**
  * @author Administrator
- *         2017/4/20 0020
+ * 2017/4/20 0020
  */
 public class FileSizeUtil {
     /**
